♻️ WV-18 refactor : api가 없어 경기일정 임의로 작성한 것 분리
