+++ components/main/schedule.tsx
@@ -9,25 +9,9 @@ import TeamRanking from "@/assets/images/team_ranking.png";
 import ParkingReservaion from "@/assets/images/parking_reservation.png";
 import IconButton from "@/components/ui/icon-button";
 import { ChevronLeft, ChevronRight } from "lucide-react";
+import { SCHEDULE_LIST } from "@/constants/schedules";
 
 type DayKey = keyof typeof SCHEDULE_LIST;
-
-export const SCHEDULE_LIST = {
-  DAY1: {
-    date: "2024.10.11",
-    placeTime: "잠실 18:00",
-    score: "1",
-    target: "LG",
-    targetScore: "4",
-  },
-  DAY2: {
-    date: "2024.10.08",
-    placeTime: "잠실 18:00",
-    score: "3",
-    target: "KIA",
-    targetScore: "2",
-  },
-} as const;
 
 const Schedule = () => {
   const [scheduleData, setScheduleData] = useState<
