+++ constants/schedules.ts
@@ -1,0 +1,16 @@
+export const SCHEDULE_LIST = {
+  DAY1: {
+    date: "2024.10.11",
+    placeTime: "잠실 18:00",
+    score: "1",
+    target: "LG",
+    targetScore: "4",
+  },
+  DAY2: {
+    date: "2024.10.08",
+    placeTime: "잠실 18:00",
+    score: "3",
+    target: "KIA",
+    targetScore: "2",
+  },
+} as const;
